fix: 중복 폴더 삭제
