+++ src/pages/Review/components/Card/ReviewPostPreview/ReviewPostPreview.container.js
@@ -1,5 +1,5 @@
 import ReviewPostPreviewUI from './ReviewPostPreview.presenter';
 
-export default function ReviewPostPreview() {
-  return <ReviewPostPreviewUI />;
+export default function ReviewPostPreview({ isHovered, setIsHovered }) {
+  return <ReviewPostPreviewUI isHovered={isHovered} setIsHovered={setIsHovered} />;
 }
