--- src/pages/Review/components/ReviewContent/ReviewPostFilter/ReviewPostFilter.styles.js
@@ -1,18 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  width: 100%;
-  margin: 0 auto;
-  border: 3px solid #aaa;
-  display: flex;
-  justify-content: space-between;
-  padding: 0 10%;
-`;
-
-export const ContentBox = styled.div`
-  width: 25%;
-`;
-
-export const AlignBox = styled.div`
-  width: 25%;
-`;
