--- src/pages/Review/components/ReviewContent/ReviewPostPreview/ReviewPostPreview.container.js
@@ -1,7 +1,0 @@
-import ReviewPostPreviewUI from './ReviewPostPreview.presenter';
-
-function ReviewPostPreview({ expanded, setIsExpanded }) {
-  return <ReviewPostPreviewUI expanded={expanded} setIsExpanded={setIsExpanded} />;
-}
-
-export default ReviewPostPreview;
