--- src/pages/Review/components/ReviewContent/ReviewPostPreview/ReviewPostPreview.styles.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const Box = styled.div`
-  height: 10rem;
-  width: 10rem;
-  background-color: rgba(0, 0, 0, 0.56);
-  z-index: 99999;
-  position: absolute;
-  transform: translateY(-300px);
-`;
